MODREQMED-18 Extract interface for SearchService
